TJ:Thiago: Adicionando teste unitários e de integração. Mudando propriedade 'AguardandoRespostaDe' para usar enum e passando método para dentro da entidade. (pendente boa parte da refatoração, mas testes foram criados para garantir que refatoração não quebre o que já funciona)

+++ HomeTask.Infrastructure/Repositories/ServicoRepositories.cs
@@ -57,8 +57,8 @@ public class ServicoPrestadorRepository : RepositoryBase<ServicoPrestador>, ISer
             query = query.Where(s => s.PrecoBase <= precoMaximo.Value);
 
         return await query
-            .OrderByDescending(s => s.Prestador.MediaAvaliacoes)
-            .ThenBy(s => s.PrecoBase)
+            .OrderByDescending(s => (double)s.Prestador.MediaAvaliacoes)
+            .ThenBy(s => (double)s.PrecoBase)
             .ToListAsync(cancellationToken);
     }
